Handle basic escape sequences in values
This is more or less a hack based on information gleaned from

http://developer.apple.com/library/mac/#documentation/cocoa/conceptual/LoadingResources/Strings/Strings.html

It only escapes NLs, CRs, tabs, double quotes and backslashes.  Unicode
escapes are not handled because I don't need them right now.  In the
case of unknown characters it does not remove the backslash which I did
to be minimally invasive, however I did not confirm the actual behavior
on Apple's side.